feat: refactor constants and types
