+++ types/partner.d.ts
@@ -1,0 +1,8 @@
+export interface Partner {
+  logo: string;
+  status: string;
+  companyname: string;
+  location: string;
+  description: string;
+  stand: string;
+}
